allow the user to pass an argument to change what is said

+++ ramsay.py
@@ -1,21 +1,27 @@
+from sys import argv
 import random
-xlist = [
-    "My gran could do better... AND SHE'S DEAD.", \
-    "for what we are about to eat, may the lord make us truly not vomit.", \
-    "this lamb is so undercooked, it's following mary to school!", \
-    "this pizza is so disgusting, if you take it to italy you'll get arrested.", \
-    "this steak is so burnt, luke thought it was his uncle owen.", \
-    "you used so much oil, the U.S. wants to invade the plate.", \
-    "your food is so bad even a canadian would insult you.", \
-    "this fish is so raw he's still finding nemo", \
-    "this isn’t a pizza, this is a mistake. this is an italian tragedy.", \
-    "there’s enough garlic in here to kill every vampire in rurope.", \
-    "hey, panini head, are you even listening to me?", \
-    "i grew up in a funny way.", \
-    "this crab is so undercooked i can still hear it singing ‘under the sea!’", \
-    "i dress for women and i undress for men."
-]
-x = random.choice(xlist)
+
+if len(argv) < 2:
+    xlist = [
+        "My gran could do better... AND SHE'S DEAD.", \
+        "for what we are about to eat, may the lord make us truly not vomit.", \
+        "this lamb is so undercooked, it's following mary to school!", \
+        "this pizza is so disgusting, if you take it to italy you'll get arrested.", \
+        "this steak is so burnt, luke thought it was his uncle owen.", \
+        "you used so much oil, the U.S. wants to invade the plate.", \
+        "your food is so bad even a canadian would insult you.", \
+        "this fish is so raw he's still finding nemo", \
+        "this isn’t a pizza, this is a mistake. this is an italian tragedy.", \
+        "there’s enough garlic in here to kill every vampire in rurope.", \
+        "hey, panini head, are you even listening to me?", \
+        "i grew up in a funny way.", \
+        "this crab is so undercooked i can still hear it singing ‘under the sea!’", \
+        "i dress for women and i undress for men."
+    ]
+    x = random.choice(xlist)
+else:
+    x = ' '.join(argv[1:])
+
 dashdash = "-" * len(x)
 print("                               ")
 print("⠁⡼⠋⠀⣆⠀⠀⣰⣿⣫⣾⢿⣿⣿⠍⢠⠠⠀⠀⢀⠰⢾⣺⣻⣿⣿⣿⣷⡀⠀ ")
